update manajemen kategori

+++ PasarLocal/routes/web.php
@@ -12,9 +12,10 @@ Route::post('/register', [AuthController::class, 'register']);
 Route::post('/login', [AuthController::class, 'login']);
 Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth');
 
-Route::get('/kategori-produk', [KategoriProdukController::class, 'index'])->name('kategori-produk.index');
-Route::get('/kategori-produk/create', [KategoriProdukController::class, 'create'])->name('kategori-produk.create');
-Route::post('/kategori-produk', [KategoriProdukController::class, 'store'])->name('kategori-produk.store');
-Route::get('/kategori-produk/{id}/edit', [KategoriProdukController::class, 'edit'])->name('kategori-produk.edit');
-Route::put('/kategori-produk/{id}', [KategoriProdukController::class, 'update'])->name('kategori-produk.update');
-Route::delete('/kategori-produk/{id}', [KategoriProdukController::class, 'destroy'])->name('kategori-produk.destroy');
+#kategori-produk
+Route::get('/admin/kategori-produk', [KategoriProdukController::class, 'index'])->name('admin.kategori-produk.index');
+Route::get('/admin/kategori-produk/create', [KategoriProdukController::class, 'create'])->name('admin.kategori-produk.create');
+Route::post('/admin/kategori-produk', [KategoriProdukController::class, 'store'])->name('admin.kategori-produk.store');
+Route::get('/admin/kategori-produk/{id}/edit', [KategoriProdukController::class, 'edit'])->name('admin.kategori-produk.edit');
+Route::put('/admin/kategori-produk/{id}', [KategoriProdukController::class, 'update'])->name('admin.kategori-produk.update');
+Route::delete('/admin/kategori-produk/{id}', [KategoriProdukController::class, 'destroy'])->name('admin.kategori-produk.destroy');
